+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/components/branding/options.ts
@@ -413,4 +413,22 @@ export const options = [
     logoTitleEn: "Department of Justice Canada",
     logoTitleFr: "Ministère de la Justice Canada",
   },
+  {
+    name: "nserc",
+    urlEn: "https://nserc-crsng.canada.ca/en",
+    urlFr: "https://nserc-crsng.canada.ca/fr",
+    logoEn: "/img/branding/nserc-en.svg",
+    logoFr: "/img/branding/nserc-fr.svg",
+    logoTitleEn: "Natural Sciences and Engineering Research Council of Canada",
+    logoTitleFr: "Conseil de recherches en sciences naturelles et en génie du Canada",
+  },
+  {
+    name: "agriculture",
+    urlEn: "https://agriculture.canada.ca/en",
+    urlFr: "https://agriculture.canada.ca/fr",
+    logoEn: "/img/branding/agriculture-en.svg",
+    logoFr: "/img/branding/agriculture-fr.svg",
+    logoTitleEn: "Agriculture and Agri-Food Canada",
+    logoTitleFr: "Agriculture et Agroalimentaire Canada",
+  },
 ];