+++ next.config.ts
@@ -4,6 +4,7 @@ const nextConfig: NextConfig = {
   reactCompiler: {
     compilationMode: "annotation",
   },
+  cacheComponents: false,
   sassOptions: {
     includePaths: ["./styles"],
   },
@@ -69,7 +70,6 @@ const nextConfig: NextConfig = {
       // Note: this matches values in constants.ts
       bodySizeLimit: "10mb",
     },
-    viewTransition: true,
   },
   typedRoutes: true,
   turbopack: {