+++ packages/connectors/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2026-08-04
+
+- Pass error reason through Error message instead of using a custom `ErrorWithCause` class
+
 ## [3.0.0] - 2026-07-02
 
 - Rework `GCNotifyConnector` and `Notification` modules to prepare for new notification pipeline